Add files via upload
I was given an assignment to create an article on Next.js for my 2nd class. I wasn't able to make it perfect, but I tried my best to complete it. Thanks to Mam Hina Naseer Ahmed for assigning it and covering HTML with us. Thank you so much! I hope you like my article.